[Refactor] refactor iceberg metadata scanner
