+++ fe/fe-core/src/main/java/com/starrocks/connector/metadata/iceberg/IcebergRefsTable.java
@@ -14,13 +14,11 @@
 
 package com.starrocks.connector.metadata.iceberg;
 
-import com.google.common.base.Joiner;
 import com.starrocks.analysis.DescriptorTable;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.PrimitiveType;
 import com.starrocks.catalog.ScalarType;
 import com.starrocks.catalog.Table;
-import com.starrocks.connector.ColumnTypeConverter;
 import com.starrocks.connector.ConnectorTableId;
 import com.starrocks.connector.metadata.MetadataTable;
 import com.starrocks.connector.metadata.MetadataTableType;
@@ -29,7 +27,6 @@ import com.starrocks.thrift.TTableDescriptor;
 import com.starrocks.thrift.TTableType;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static com.starrocks.connector.metadata.TableMetaMetadata.METADATA_DB_NAME;
 
@@ -63,21 +60,7 @@ public class IcebergRefsTable extends MetadataTable {
     public TTableDescriptor toThrift(List<DescriptorTable.ReferencedPartitionInfo> partitions) {
         TTableDescriptor tTableDescriptor = new TTableDescriptor(getId(), TTableType.ICEBERG_REFS_TABLE,
                 fullSchema.size(), 0, getName(), METADATA_DB_NAME);
-        THdfsTable hdfsTable = new THdfsTable();
-
-        hdfsTable.setColumns(fullSchema.stream().map(Column::toThrift).collect(Collectors.toList()));
-        hdfsTable.setPartition_columnsIsSet(false);
-
-        String columnNames = Joiner.on(',').join(fullSchema.stream()
-                .map(Column::getName)
-                .collect(Collectors.toList()));
-        hdfsTable.setHive_column_names(columnNames);
-
-        String columnTypes = Joiner.on(',').join(fullSchema.stream()
-                .map(x -> ColumnTypeConverter.toHiveType(x.getType()))
-                .collect(Collectors.toList()));
-        hdfsTable.setHive_column_types(columnTypes);
-
+        THdfsTable hdfsTable = buildThriftTable(fullSchema);
         tTableDescriptor.setHdfsTable(hdfsTable);
         return tTableDescriptor;
     }
