+++ fe/fe-core/src/main/java/com/starrocks/connector/metadata/iceberg/LogicalIcebergMetadataTable.java
@@ -14,13 +14,11 @@
 
 package com.starrocks.connector.metadata.iceberg;
 
-import com.google.common.base.Joiner;
 import com.starrocks.analysis.DescriptorTable;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.PrimitiveType;
 import com.starrocks.catalog.ScalarType;
 import com.starrocks.catalog.Table;
-import com.starrocks.connector.ColumnTypeConverter;
 import com.starrocks.connector.ConnectorTableId;
 import com.starrocks.connector.metadata.MetadataTable;
 import com.starrocks.connector.metadata.MetadataTableType;
@@ -75,23 +73,11 @@ public class LogicalIcebergMetadataTable extends MetadataTable {
     public TTableDescriptor toThrift(List<DescriptorTable.ReferencedPartitionInfo> partitions) {
         TTableDescriptor tTableDescriptor = new TTableDescriptor(getId(), TTableType.LOGICAL_ICEBERG_METADATA_TABLE,
                 fullSchema.size() - PLACEHOLDER_COLUMNS.size(), 0, getName(), METADATA_DB_NAME);
-        THdfsTable hdfsTable = new THdfsTable();
         List<Column> columns = fullSchema.stream()
                 .filter(c -> !PLACEHOLDER_COLUMNS.contains(c))
                 .collect(Collectors.toList());
 
-        hdfsTable.setColumns(columns.stream().map(Column::toThrift).collect(Collectors.toList()));
-        hdfsTable.setPartition_columnsIsSet(false);
-
-        String columnNames = Joiner.on(',').join(columns.stream()
-                .map(Column::getName)
-                .collect(Collectors.toList()));
-        hdfsTable.setHive_column_names(columnNames);
-
-        String columnTypes = Joiner.on(',').join(columns.stream()
-                .map(x -> ColumnTypeConverter.toHiveType(x.getType()))
-                .collect(Collectors.toList()));
-        hdfsTable.setHive_column_types(columnTypes);
+        THdfsTable hdfsTable = buildThriftTable(columns);
         tTableDescriptor.setHdfsTable(hdfsTable);
         return tTableDescriptor;
     }
