+++ java-extensions/iceberg-metadata-reader/src/main/java/com/starrocks/connector/iceberg/IcebergHistoryTableScanner.java
@@ -15,68 +15,35 @@
 package com.starrocks.connector.iceberg;
 
 import com.google.common.collect.ImmutableSet;
-import com.starrocks.jni.connector.ColumnType;
 import com.starrocks.jni.connector.ColumnValue;
-import com.starrocks.jni.connector.ConnectorScanner;
 import com.starrocks.utils.loader.ThreadContextClassLoader;
 import org.apache.iceberg.Snapshot;
-import org.apache.iceberg.Table;
 import org.apache.iceberg.util.SnapshotUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
-import java.util.TimeZone;
 
-import static org.apache.iceberg.util.SerializationUtil.deserializeFromBase64;
-
-public class IcebergHistoryTableScanner extends ConnectorScanner {
+public class IcebergHistoryTableScanner extends AbstractIcebergMetadataScanner {
     private static final Logger LOG = LogManager.getLogger(IcebergHistoryTableScanner.class);
 
-    private final String serializedTable;
-    private final String[] requiredFields;
-    private final String[] metadataColumnNames;
-    private final String[] metadataColumnTypes;
-    private ColumnType[] requiredTypes;
-    private final int fetchSize;
-    private final ClassLoader classLoader;
-    private Table table;
-    private String timezone;
     private Iterator<Snapshot> reader;
     private Set<Long> ancestorIds;
 
     public IcebergHistoryTableScanner(int fetchSize, Map<String, String> params) {
-        this.fetchSize = fetchSize;
-        this.requiredFields = params.get("required_fields").split(",");
-        this.metadataColumnNames = params.get("metadata_column_names").split(",");
-        this.metadataColumnTypes = params.get("metadata_column_types").split(",");
-        this.serializedTable = params.get("serialized_table");
-        this.timezone = params.getOrDefault("time_zone", TimeZone.getDefault().getID());
-        this.classLoader = this.getClass().getClassLoader();
+        super(fetchSize, params);
     }
 
     @Override
-    public void open() throws IOException {
-        try (ThreadContextClassLoader ignored = new ThreadContextClassLoader(classLoader)) {
-            this.table = deserializeFromBase64(serializedTable);
-            this.ancestorIds = ImmutableSet.copyOf(SnapshotUtil.currentAncestorIds(table));
-            parseRequiredTypes();
-            initOffHeapTableWriter(requiredTypes, requiredFields, fetchSize);
-            initReader();
-        } catch (Exception e) {
-            close();
-            String msg = "Failed to open the iceberg history table reader.";
-            LOG.error(msg, e);
-            throw new IOException(msg, e);
-        }
+    public void doOpen() {
+        this.ancestorIds = ImmutableSet.copyOf(SnapshotUtil.currentAncestorIds(table));
     }
 
     @Override
-    public void close() throws IOException {
+    public void close() {
         if (reader != null) {
             reader = null;
         }
@@ -109,7 +76,8 @@ public class IcebergHistoryTableScanner extends ConnectorScanner {
         }
     }
 
-    private void initReader() {
+    @Override
+    protected void initReader() {
         reader = table.snapshots().iterator();
     }
 
@@ -125,19 +93,6 @@ public class IcebergHistoryTableScanner extends ConnectorScanner {
                 return ancestorIds.contains(snapshot.snapshotId()) ? "true" : "false";
             default:
                 throw new IllegalArgumentException("Unrecognized column name " + columnName);
-        }
-    }
-
-    private void parseRequiredTypes() {
-        HashMap<String, String> columnNameToType = new HashMap<>();
-        for (int i = 0; i < metadataColumnNames.length; i++) {
-            columnNameToType.put(metadataColumnNames[i], metadataColumnTypes[i]);
-        }
-
-        requiredTypes = new ColumnType[requiredFields.length];
-        for (int i = 0; i < requiredFields.length; i++) {
-            String type = columnNameToType.get(requiredFields[i]);
-            requiredTypes[i] = new ColumnType(requiredFields[i], type);
         }
     }
 }
