+++ java-extensions/iceberg-metadata-reader/src/main/java/com/starrocks/connector/iceberg/IcebergMetadataScanner.java
@@ -19,9 +19,7 @@ import com.esotericsoftware.kryo.io.Output;
 import com.google.common.collect.ImmutableList;
 import com.starrocks.connector.share.iceberg.CommonMetadataBean;
 import com.starrocks.connector.share.iceberg.IcebergMetricsBean;
-import com.starrocks.jni.connector.ColumnType;
 import com.starrocks.jni.connector.ColumnValue;
-import com.starrocks.jni.connector.ConnectorScanner;
 import com.starrocks.utils.loader.ThreadContextClassLoader;
 import de.javakaffee.kryoserializers.UnmodifiableCollectionsSerializer;
 import org.apache.iceberg.ContentFile;
@@ -31,7 +29,6 @@ import org.apache.iceberg.ManifestFile;
 import org.apache.iceberg.ManifestFiles;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.StructLike;
-import org.apache.iceberg.Table;
 import org.apache.iceberg.expressions.Expression;
 import org.apache.iceberg.expressions.Expressions;
 import org.apache.iceberg.io.CloseableIterator;
@@ -41,7 +38,6 @@ import org.apache.logging.log4j.Logger;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -49,7 +45,7 @@ import java.util.stream.Collectors;
 import static org.apache.iceberg.util.ByteBuffers.toByteArray;
 import static org.apache.iceberg.util.SerializationUtil.deserializeFromBase64;
 
-public class IcebergMetadataScanner extends ConnectorScanner {
+public class IcebergMetadataScanner extends AbstractIcebergMetadataScanner {
     private static final Logger LOG = LogManager.getLogger(IcebergMetadataScanner.class);
 
     protected static final List<String> SCAN_COLUMNS =
@@ -98,15 +94,7 @@ public class IcebergMetadataScanner extends ConnectorScanner {
     private static final int DELETE_FILE = 1;
     private final String manifestBean;
     private final String predicateInfo;
-    private final String serializedTable;
-    private final String[] requiredFields;
-    private final String[] metadataColumnNames;
-    private final String[] metadataColumnTypes;
-    private ColumnType[] requiredTypes;
-    private final int fetchSize;
-    private final ClassLoader classLoader;
-    private Table table;
-    private boolean loadColumnStats;
+    private final boolean loadColumnStats;
     private Expression predicate;
     private ManifestFile manifestFile;
     private Kryo kryo;
@@ -115,33 +103,17 @@ public class IcebergMetadataScanner extends ConnectorScanner {
     private CloseableIterator<? extends ContentFile<?>> reader;
 
     public IcebergMetadataScanner(int fetchSize, Map<String, String> params) {
-        this.fetchSize = fetchSize;
-        this.requiredFields = params.get("required_fields").split(",");
-        this.metadataColumnNames = params.get("metadata_column_names").split(",");
-        this.metadataColumnTypes = params.get("metadata_column_types").split(",");
+        super(fetchSize, params);
         this.predicateInfo = params.get("serialized_predicate");
-        this.serializedTable = params.get("serialized_table");
         this.manifestBean = params.get("split_info");
         this.loadColumnStats = Boolean.parseBoolean(params.get("load_column_stats"));
-        this.classLoader = this.getClass().getClassLoader();
     }
 
     @Override
-    public void open() throws IOException {
-        try (ThreadContextClassLoader ignored = new ThreadContextClassLoader(classLoader)) {
-            this.predicate = predicateInfo.isEmpty() ? Expressions.alwaysTrue() : deserializeFromBase64(predicateInfo);
-            this.manifestFile = deserializeFromBase64(manifestBean);
-            this.table = deserializeFromBase64(serializedTable);
-            parseRequiredTypes();
-            initOffHeapTableWriter(requiredTypes, requiredFields, fetchSize);
-            initReader();
-            initSerializer();
-        } catch (Exception e) {
-            close();
-            String msg = "Failed to open the iceberg metadata reader.";
-            LOG.error(msg, e);
-            throw new IOException(msg, e);
-        }
+    public void doOpen() {
+        this.predicate = predicateInfo.isEmpty() ? Expressions.alwaysTrue() : deserializeFromBase64(predicateInfo);
+        this.manifestFile = deserializeFromBase64(manifestBean);
+        initSerializer();
     }
 
     @Override
@@ -186,7 +158,8 @@ public class IcebergMetadataScanner extends ConnectorScanner {
         }
     }
 
-    private void initReader() {
+    @Override
+    protected void initReader() {
         Map<Integer, PartitionSpec> specs = table.specs();
         List<String> scanColumns;
         if (manifestFile.content() == ManifestContent.DATA) {
@@ -303,18 +276,5 @@ public class IcebergMetadataScanner extends ConnectorScanner {
                 .collect(Collectors.toMap(
                         Map.Entry::getKey,
                         entry -> toByteArray(entry.getValue())));
-    }
-
-    private void parseRequiredTypes() {
-        HashMap<String, String> columnNameToType = new HashMap<>();
-        for (int i = 0; i < metadataColumnNames.length; i++) {
-            columnNameToType.put(metadataColumnNames[i], metadataColumnTypes[i]);
-        }
-
-        requiredTypes = new ColumnType[requiredFields.length];
-        for (int i = 0; i < requiredFields.length; i++) {
-            String type = columnNameToType.get(requiredFields[i]);
-            requiredTypes[i] = new ColumnType(requiredFields[i], type);
-        }
     }
 }
